cleanup font configs
