+++ packages/asset-uploader/package.json
@@ -16,6 +16,7 @@
     "@aws-sdk/client-s3": "^3.131.0",
     "@aws-sdk/lib-storage": "^3.131.0",
     "@webstudio-is/prisma-client": "*",
+    "@webstudio-is/fonts": "*",
     "app-root-path": "^3.0.0",
     "fontkit": "^2.0.2",
     "sharp": "^0.30.7"
