+++ packages/asset-uploader/src/constants.ts
@@ -1,2 +1,1 @@
 export const DEFAULT_UPLOAD_PATH = "public/uploads";
-export const FONT_FORMATS = ["ttf", "woff", "woff2", "otf"] as const;
