+++ packages/asset-uploader/src/utils/filter-by-type.ts
@@ -1,12 +1,12 @@
+import { FONT_FORMATS } from "@webstudio-is/fonts";
 import type { AssetType } from "../types";
-import { FONT_FORMATS } from "../constants";
 
 export const filterByType = <PartialAsset extends { format: string }>(
   assets: Array<PartialAsset>,
   type: AssetType
 ) => {
   return assets.filter(({ format }: PartialAsset) => {
-    const isFont = FONT_FORMATS.includes(format as typeof FONT_FORMATS[number]);
+    const isFont = format in FONT_FORMATS;
     if (type === "font") {
       return isFont;
     }
