+++ packages/asset-uploader/src/utils/format-asset.ts
@@ -1,8 +1,8 @@
 import { z } from "zod";
 import { Asset as DbAsset } from "@webstudio-is/prisma-client";
+import { type FontFormat, FONT_FORMATS } from "@webstudio-is/fonts";
 import { Asset } from "../types";
 import { getAssetPath } from "./get-asset-path";
-import { FONT_FORMATS } from "../constants";
 import { styles } from "./get-font-data";
 
 const ImageMeta = z.object({
@@ -21,13 +21,12 @@ export type FontMeta = z.infer<typeof FontMeta>;
 export const formatAsset = (asset: DbAsset): Asset => {
   const base = { ...asset, path: getAssetPath(asset) };
 
-  const isFont = FONT_FORMATS.includes(
-    asset.format as typeof FONT_FORMATS[number]
-  );
+  const isFont = asset.format in FONT_FORMATS;
 
   if (isFont) {
     return {
       ...base,
+      format: asset.format as FontFormat,
       meta: FontMeta.parse(JSON.parse(asset.meta)),
     };
   }
