+++ packages/fonts/src/types.ts
@@ -1,0 +1,1 @@
+export type FontFormat = "ttf" | "woff" | "woff2" | "otf";
